+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### v1.0.8
+### v1.0.9
 - Feature
   - Stack trace duplicates logged by hash reference
   - variables now have 2 separate properties
@@ -14,6 +14,7 @@
   - Removed tracers from configuration will have their classes restored to the original
   - Properly handle recursion if a low level method is traced that is used by the Tracer Agent itself
   - line # is no longer required.  Will default to 0, which is first line of method when not specified.
+  - regex matches for variable values now uses find so that in effect the exact match now becomes a contains test
 
 ### v0.5.1
 - Initial release for general use